Restored call to enemy_attack()
Ideally this call will be replaced with a signal at some point but this works.

The player now dies when they reach zero health.

As of now there is no death animation.
Adding a death animation would require adding code to pause player movement, drawing and playing a death animation, and then emitting the player death signal when the animation is over.