Level Backlinks
Levels can now link to the previous level.
When a previous level is loaded thee player is teleported to the location of the EndLocator node.
This node should be positioned just before the transition to the next level so that the player does not go from the beginning of level 2 to the beginning of level one. Instead the player will go from the beginning of level two to the end of level one.