Resolved error from double-jump activation
Yeoungjun had added code to the player script which depended on a node which only existed in one level. This caused the game to throw an error while loading any other level.

Removed this method from player script and moved the functionality to a script in the node which detects the end of the level. Animation now plays with no errors.

Need to work out why double jump is always active.

Co-Authored-By: godofocean7 <171698278+godofocean7@users.noreply.github.com>