Refactored level three to template
Yeoungjun didn't create level three according to procedure so it was not
compatible with the game backend.
He said it was refactored to inherit the template but it still wasn't
quite right so I copied all the contents into a new level inheriting the
template.

Also added check to signal auto linking that prevents linking to
previously linked signals.