Fixed broken reference and orphaned script
The blood splash had two scripts, blood_splash.gd and node_2d.gd.
Only the latter was being used. node_2d.gd is an awful name so I renamed
it to blood_splash.gd and removed the unused blood_splash.gd.

The game was throwing errors when loading levels with ribbugs as the
script was attempting to reference a nonexistant area2d node so I
removed this unused reference.

Also renamed the level one root node.